Refactor into files and fix replacement paths
